Changes to the user class
Created separate sub-packages for books  and users in every package. Changed the user @Id to be the username as it is unique and not null and there is no need for a separate numeric id. Added a first and last name to the user details. Added length constraints for the username. Created a user dto.